Improves ProcessRunner to respect memory_limit in PHP scripts
When running scripts we now check if they are PHP scripts, we assert
here that PHP scripts should start "php ". When running a PHP script we
check if the script defines the memory_limit flag, if this is not
defined we will add it using the current ini value.

Scripts that define a memory_limit will be left untouched.